New Chevalier module. Gilles e-mail:
you'll find attached a new version of the Chevalier module, which offers a
better interpolation of the hydro profiles (density, velocity, pressure).
The arrays are now searched by dichotomy instead of element by element, and
the requested value is computed with a cubic spline instead of a linear
interpolation.  This makes the three functions SNR_density(),
SNR_velocity(), and SNR_pressure() both faster and more accurate (the way
they are called is unaffected).

Accuracy is not a huge concern with Ramses though, as the space resolution
we use in 3D can easily be outperformed in 1D.  The integrator step of
Chevalier is set by SN_TECH%step in init_time.f90.  I recall that the
integration is made on the self-similarity variable eta, so that the actual
resolution in radius is only known at the end of the computation, and is non
uniform (the extrema are stored as dr_min and dr_max in the SNR structure
for info).  The default value step = 1e-3 should be ok for most cases -- you
just have to check that the profiles are over-resolved, as indicated in
Ramses output just after the call to Chevalier (the max value of dr_SNR must
me lower than the min value of dr_ramses).

Speed is a concern during the initialisation of Ramses, as building the
initial AMR grid always involves multiple passes at each level.  Using the
new version of Chevalier can reduce the time it takes by a factor of about 2
when going to level 8.  This will make a sizeable difference when developing
on your laptop.

Finally, the ionisation and radiative tracers indirectly benefit from the
changes, but they are still limited by the crude linear reconstruction in
time: the next step will be to compute the profiles history.